rag-system/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
openai==1.52.0
qdrant-client==1.12.1
sqlmodel==0.0.22
instructor==1.6.2
python-dotenv==1.0.1
tenacity==9.0.0
llama-index==0.11.12
llama-index-embeddings-openai==0.2.0
llama-index-llms-openai==0.3.0
sentence-transformers==3.3.0
markdown==3.7
beautifulsoup4==4.12.3
pydantic==2.9.2
python-multipart==0.0.20
aiofiles==24.1.0
requests==2.32.3